webcrypto: Move encryption package out of task
