+++ task/import.go
@@ -18,7 +18,7 @@ import (
 	"github.com/livepeer/go-api-client/logs"
 	"github.com/livepeer/go-tools/drivers"
 	"github.com/livepeer/livepeer-data/pkg/data"
-	"github.com/livepeer/task-runner/task/encryption"
+	"github.com/livepeer/task-runner/webcrypto"
 )
 
 const IPFS_PREFIX = "ipfs://"
@@ -102,7 +102,7 @@ func getFile(ctx context.Context, osSess drivers.OSSession, cfg ImportTaskConfig
 	switch params.Encryption.Algorithm {
 	case "", "aes-256-cbc":
 		glog.V(logs.VVERBOSE).Infof("Decrypting file with key file=%s keyHash=%x", params.URL, sha256.Sum256([]byte(params.Encryption.Key)))
-		decrypted, err := encryption.DecryptAES256CBCReader(content, params.Encryption.Key)
+		decrypted, err := webcrypto.DecryptAESCBC(content, params.Encryption.Key)
 		if err != nil {
 			content.Close()
 			return "", 0, nil, fmt.Errorf("failed to decrypt input file: %w", err)
